Reduced precision of binary angle measurements to 16 bit

+++ src/r_state.h
@@ -118,7 +118,7 @@ extern angle_t		rw_normalangle;
 
 
 // angle to line origin
-extern int		rw_angle1;
+extern int16_t		rw_angle1;
 
 // Segs count?
 extern int		sscount;
